fix 2 issues
`s:is_unmapped()` always considered the mode in
which we were interested to be `normal`:
it could another mode.

`s:translate_special_key()` failed to translate
a key such as `yCTRL-`.
Because of this, `s:is_unmapped()` thought that
`yCTRL-` was not free because we have a mapping
beginning with `y`.
Long story short, when `s:translate_special_key()`
receives a key ending with `CTRL-`, it should
return an empty string, so that `mapcheck()`
doesn't return anything.